Add hooks and script for installing them
Add hooks for making sure commit messages follow a given template,
c-files are formatted correctly, for making sure branch naming
convention is followed and for making sure the filename is
in ascii. For more information about the hooks, see CT-22 where most
of the functionality of them was implemented and documented. Also
added a check on the prepare-commit-message hook to check if the
correct commit message template was used. The commit message template
file was renamed to ensure better readability.

Add test project to make sure the hooks are installed in
submodules also. This also served as a test to see that submodules
work.

Add script for installing hooks in both the main repo and all
subrepos.

Add clang-format file in order to get all braces on newlines.

Files added:
- .clang-format: Config file used for clang
- .gitmodules: Add test submodule
- hooks/.commit-msg-template: Change name for readability
- hooks/.users: Authorized users to use to check branch naming
- hooks/.branch-categories: List of branch categories to use to
check branch naming
- hooks/commit-msg: Add hook
- hooks/pre-commit: Add hook
- hooks/prepare-commit-msg: Add hook and now check if a valid
template is used
- projects/git-submodule-test: Add test repo
- scripts/install_hooks: Add script for installing hooks in all repos

CT-24