Adding nrfjprog script
Since we will be flashing and communicating with the development kits
through a raspberry pi ubuntu server, the regular nrfjprog command
line tools will not work (does not work for arm arch). This script
will serve as a replacement and make it possible to perform the
most important nrfjprog actions.

CT-45