Add ceedling project file and pipeline file
Make a new folder tdd/, an asset folder for tdd-related files. Add
ceedling related files project.yml and bitbucket-pipelines.yml. If we
use different testing tools in the future, we can add a new folder for
that inside the tdd/ folder similar to how it has been done with
ceedling here.

These files are only assets for the project repositories to use and
will not be used in the main repo.

Note that the bitbucket-pipelines.yml file is specific
to ceedling hence is why it is inside the ceedling/ folder.

Documentation for ceedling and the pipeline can be found on the wiki
under "Test driven development"

Files added:
- tdd/ceedling/bitbucket-pipelines.yml: Specifies what will be run
in the bitbucket pipeline
- tdd/ceedling/project.yml: Ceedling config file set up for our
project

CT-34