Fix compiler path and nrfjprog bug
Arm GCC compiler is not found without making changes to the
Makefile.posix in the nRF SDK. This fixes the issue. I have just edited
the path on where to find the Arm GCC compiler. It finds the version of
the compiler that you have on the computer, so you don't need the newest
version of the compiler.

I also had to fix a bug in the nrfjprog.sh script. Flashing the
softdevice was not working. This was just a minor mistake. I only had to
add a couple of spaces. I also added some more description to print when
you use the script incorrectly.

Files modified:
- scripts/nrfjprog.sh: Fixing minor logic error to flashing softdevice,
and added some more prints.
- shared_lib/external/nRF5_SDK_17.0.0/components/toolchain/gcc/Makefile.posix: Changed the path on where to find the Arm GCC compiler.

CT-29