Fix bug with branch naming and hook installation
Fix install_hooks so that it copies over all the hidden files also
(.users, .branch-categories, .commit-msg-template). This was not
striclty speak necessary, as I also changed the pre-commit hook to
use these files from the main repo's hooks/ folder. Still nice for
comepleteness.

The pre-commit hook now looks for users and branch categories in the
right place.

Updated requirements to include that it is necessary to install
clang-format

Files modified:
- README.md: Update requirements to include clang-format
- hooks/pre-commit: Fix bug locating .users and .branch-categories
files
- scripts/install_hooks: Make script copy hidden files also

CT-51, CT-52